+++ src/pybel/testing/cases.py
@@ -30,7 +30,7 @@ class TemporaryCacheMixin(unittest.TestCase):
             self.connection = 'sqlite:///' + self.path
             log.info('Test generated connection string %s', self.connection)
 
-        self.manager = Manager(connection=self.connection)
+        self.manager = Manager.from_connection(connection=self.connection)
         self.manager.create_all()
 
     def tearDown(self):
@@ -57,7 +57,7 @@ class TemporaryCacheClsMixin(unittest.TestCase):
             cls.connection = 'sqlite:///' + cls.path
             log.info('Test generated connection string %s', cls.connection)
 
-        cls.manager = Manager(connection=cls.connection)
+        cls.manager = Manager.from_connection(cls.connection)
         cls.manager.create_all()
 
     @classmethod